Initial draft of benchmarks
