+++ build.yaml
@@ -4005,6 +4005,28 @@ targets:
   - linux
   - posix
   uses_polling: false
+- name: bm_filters
+  build: test
+  language: c++
+  src:
+  - test/cpp/microbenchmarks/bm_filters.cc
+  deps:
+  - grpc_benchmark
+  - benchmark
+  - grpc++_test_util_unsecure
+  - grpc_test_util_unsecure
+  - grpc++_unsecure
+  - grpc_unsecure
+  - gpr_test_util
+  - gpr
+  - grpc++_test_config
+  benchmark: true
+  defaults: benchmark
+  platforms:
+  - mac
+  - linux
+  - posix
+  uses_polling: false
 - name: bm_fullstack_streaming_ping_pong
   build: test
   language: c++
