+++ test/cpp/microbenchmarks/noop-benchmark.cc
@@ -22,7 +22,9 @@
 #include <benchmark/benchmark.h>
 
 static void BM_NoOp(benchmark::State& state) {
+  int i = 0;
   while (state.KeepRunning()) {
+    i++;
   }
 }
 BENCHMARK(BM_NoOp);
